refactor: filterconnections, taxonomy components folder
